requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==7.0.1
cryptography==38.0.4
Deprecated==1.2.13
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
drf-writable-nested==0.7.0
dry-rest-permissions==0.1.10
ecdsa==0.18.0
environ==1.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
future==0.18.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
psycopg2==2.9.5
pyasn1==0.4.8
rsa==4.9
six==1.16.0
sqlparse==0.4.3
swagger-ui==0.1.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1